+++ packages/create-robo/src/index.ts
@@ -20,6 +20,7 @@ interface CommandOptions {
 	template?: string
 	typescript?: boolean
 	verbose?: boolean
+	roboVersion?: string
 }
 
 new Command('create-robo <projectName>')
@@ -32,6 +33,7 @@ new Command('create-robo <projectName>')
 	.option('-t --template <templateUrl>', 'create a Robo from an online template')
 	.option('-ts --typescript', 'create a Robo using TypeScript')
 	.option('-v --verbose', 'print more information for debugging')
+	.option('-rv, --robo-version <value>', 'choose which version of robo your project will use')
 	.action(async (options: CommandOptions, { args }) => {
 		logger({
 			level: options.verbose ? 'debug' : 'info'
@@ -40,6 +42,9 @@ new Command('create-robo <projectName>')
 		logger.debug(`Package manager:`, getPackageManager())
 		logger.debug(`create-robo version:`, packageJson.version)
 		logger.debug(`Current working directory:`, process.cwd())
+
+		// parses robo version argument
+		const rv = await getRoboversionArg(options.roboVersion)
 
 		// Check for updates
 		await checkUpdates()
@@ -104,7 +109,7 @@ new Command('create-robo <projectName>')
 
 			// Get user input to determine which features to include or use the recommended defaults
 			selectedFeaturesOrDefaults = options.features?.split(',') ?? (await robo.getUserInput())
-			await robo.createPackage(selectedFeaturesOrDefaults, options.install ?? true)
+			await robo.createPackage(selectedFeaturesOrDefaults, options.install ?? true, rv)
 
 			// Determine if TypeScript is selected and copy the corresponding template files
 			logger.debug(`Copying template files...`)
@@ -148,4 +153,20 @@ async function checkUpdates() {
 		logger.info(chalk.bold.green(`A new version of ${chalk.bold('create-robo')} is available! (v${latestVersion})`))
 		logger.info(`Run as ${chalk.bold(command)} instead to get the latest updates.`)
 	}
+}
+
+async function getRoboversionArg(roboVersion: string): Promise<string> {
+	let roboversion = 'latest'
+
+	if (roboVersion) {
+		const response = await fetch(`https://registry.npmjs.org/@roboplay/robo.js/${roboVersion}`)
+		const version = (await response.json()).version
+
+		if (version) {
+			roboversion = version
+		} else {
+			logger().error('Invalid Robo version, falling back to latest..')
+		}
+	}
+	return roboversion
 }