+++ packages/plugin-ai/src/core/openai.ts
@@ -2,15 +2,22 @@ import { logger } from './logger.js'
 import { options as pluginOptions } from '../events/_start.js'
 
 export interface GptChatMessage {
-	content: string
+	content: GptChatMessageContent
 	function_call?: GptFunctionCall
 	name?: string
 	role: 'assistant' | 'function' | 'system' | 'user'
 }
 
+export type GptChatMessageContent = string | Array<{
+	image_url?: string
+	text?: string
+	type: 'image_url' | 'text'
+}>
+
 export interface GptChatOptions {
 	backoff?: boolean
 	functions?: GptFunction[]
+	maxTokens?: number
 	messages: GptChatMessage[]
 	model?: string
 	retries?: number
@@ -41,7 +48,7 @@ export interface GptFunctionProperty {
 }
 
 export async function chat(options: GptChatOptions) {
-	const { backoff = true, functions, messages, model = 'gpt-3.5-turbo', retries = 3 } = options
+	const { backoff = true, functions, maxTokens = 1024, messages, model = 'gpt-3.5-turbo', retries = 3 } = options
 	let retryCount = 0
 
 	if (!pluginOptions.openaiKey) {
@@ -57,7 +64,8 @@ export async function chat(options: GptChatOptions) {
 					Authorization: `Bearer ${pluginOptions.openaiKey}`
 				},
 				body: JSON.stringify({
-					functions: functions?.length ? functions : undefined,
+					functions: functions?.length && !model.includes('vision') ? functions : undefined,
+					max_tokens: maxTokens,
 					messages: messages,
 					model: model
 				})