+++ packages/sage/src/core/types.ts
@@ -1,4 +1,12 @@
 export interface PackageJson {
 	dependencies?: Record<string, string>
 	devDependencies?: Record<string, string>
+	workspaces: Array<string>
+	name: string
+	repository?: {
+		directory: string
+		type: string
+		url: string
+	}
+	version: string
 }