+++ packages/sage/src/index.ts
@@ -1,22 +1,24 @@
 #!/usr/bin/env node
 import { Command } from 'commander'
 import { createRequire } from 'node:module'
+import doctorCommand from './commands/doctor.js'
 import exportCommand from './commands/export.js'
 import importCommand from './commands/import.js'
 import upgradeCommand from './commands/upgrade.js'
 import whyCommand from './commands/why.js'
+import typescriptCommand from './commands/typescript.js'
 
 // Read the version from the package.json file
 const require = createRequire(import.meta.url)
-const packageJson = require('../package.json')
+export const packageJson = require('../package.json')
 
-// TODO:
-// - Verify codemod version before running each command
 new Command('@roboplay/sage')
 	.description('Codemod for Robo.js')
 	.version(packageJson.version)
+	.addCommand(doctorCommand)
 	.addCommand(exportCommand)
 	.addCommand(importCommand)
 	.addCommand(upgradeCommand)
 	.addCommand(whyCommand)
+	.addCommand(typescriptCommand)
 	.parse(process.argv)